+++ src/app.ts
@@ -8,11 +8,10 @@ import swaggerUi from 'swagger-ui-express';
 import { StatusCodes } from 'http-status-codes';
 
 import { CredentialController } from './controllers/credentials.js';
-import { IssuerController } from './controllers/issuer.js';
-import { AccountController } from './controllers/customer.js';
+import { AccountController } from './controllers/account.js';
 import { Authentication } from './middleware/authentication.js';
 import { Connection } from './database/connection/connection.js';
-import { RevocationController } from './controllers/revocation.js';
+import { RevocationController } from './controllers/credential-status.js';
 import { CORS_ALLOWED_ORIGINS, CORS_ERROR_MSG } from './types/constants.js';
 import { LogToWebHook } from './middleware/hook.js';
 import { Middleware } from './middleware/middleware.js';
@@ -22,6 +21,10 @@ dotenv.config();
 
 // Define Swagger file
 import swaggerDocument from './static/swagger.json' assert { type: 'json' };
+import { PresentationController } from './controllers/presentation.js';
+import { KeyController } from './controllers/key.js';
+import { DIDController } from './controllers/did.js';
+import { ResourceController } from './controllers/resource.js';
 
 let swaggerOptions = {};
 if (process.env.ENABLE_AUTHENTICATION === 'true') {
@@ -113,8 +116,8 @@ class App {
 		// presentation
 		app.post(
 			`/presentation/verify`,
-			CredentialController.presentationValidator,
-			new CredentialController().verifyPresentation
+			PresentationController.presentationVerifyValidator,
+			new PresentationController().verifyPresentation
 		);
 
 		// revocation
@@ -150,19 +153,19 @@ class App {
 		);
 
 		// Keys API
-		app.post('/key/create', new IssuerController().createKey);
-		app.get('/key/read/:kid', new IssuerController().getKey);
+		app.post('/key/create', new KeyController().createKey);
+		app.get('/key/read/:kid', new KeyController().getKey);
 
 		// DIDs API
-		app.post('/did/create', IssuerController.createValidator, new IssuerController().createDid);
-		app.post('/did/update', IssuerController.updateValidator, new IssuerController().updateDid);
-		app.post('/did/deactivate/:did', IssuerController.deactivateValidator, new IssuerController().deactivateDid);
-		app.get('/did/list', new IssuerController().getDids);
-		app.get('/did/search/:did', new IssuerController().resolveDidUrl);
+		app.post('/did/create', DIDController.createDIDValidator, new DIDController().createDid);
+		app.post('/did/update', DIDController.updateDIDValidator, new DIDController().updateDid);
+		app.post('/did/deactivate/:did', DIDController.deactivateDIDValidator, new DIDController().deactivateDid);
+		app.get('/did/list', new DIDController().getDids);
+		app.get('/did/search/:did', new DIDController().resolveDidUrl);
 
 		// Resource API
-		app.post('/resource/create/:did', IssuerController.resourceValidator, new IssuerController().createResource);
-		app.get('/resource/search/:did', new IssuerController().getResource);
+		app.post('/resource/create/:did', ResourceController.resourceValidator, new ResourceController().createResource);
+		app.get('/resource/search/:did', new ResourceController().getResource);
 
 		// Account API
 		app.get('/account', new AccountController().get);
