+++ src/controllers/key.ts
@@ -1,0 +1,114 @@
+import type { Request, Response } from 'express';
+import { StatusCodes } from 'http-status-codes';
+import { IdentityServiceStrategySetup } from '../services/identity/index.js';
+
+export class KeyController {
+
+	/**
+	 * @openapi
+	 *
+	 * /key/create:
+	 *   post:
+	 *     tags: [ Key ]
+	 *     summary: Create an identity key pair.
+	 *     description: This endpoint creates an identity key pair associated with the user's account for custodian-mode clients.
+	 *     responses:
+	 *       200:
+	 *         description: The request was successful.
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: '#/components/schemas/KeyResult'
+	 *       400:
+	 *         description: A problem with the input fields has occurred. Additional state information plus metadata may be available in the response body.
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: '#/components/schemas/InvalidRequest'
+	 *             example:
+	 *               error: InvalidRequest
+	 *       401:
+	 *         $ref: '#/components/schemas/UnauthorizedError'
+	 *       500:
+	 *         description: An internal error has occurred. Additional state information plus metadata may be available in the response body.
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: '#/components/schemas/InvalidRequest'
+	 *             example:
+	 *               error: Internal Error
+	 */
+	public async createKey(request: Request, response: Response) {
+		try {
+			const key = await new IdentityServiceStrategySetup(response.locals.customer.customerId).agent.createKey(
+				'Ed25519',
+				response.locals.customer
+			);
+			return response.status(StatusCodes.OK).json(key);
+		} catch (error) {
+			return response.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
+				error: `${error}`,
+			});
+		}
+	}
+
+	/**
+	 * @openapi
+	 *
+	 * /key/read/{kid}:
+	 *   get:
+	 *     tags: [ Key ]
+	 *     summary: Fetch an identity key pair.
+	 *     description: This endpoint fetches an identity key pair's details for a given key ID. Only the user account associated with the custodian-mode client can fetch the key pair.
+	 *     parameters:
+	 *       - name: kid
+	 *         description: Key ID of the identity key pair to fetch.
+	 *         in: path
+	 *         schema:
+	 *           type: string
+	 *         required: true
+	 *     responses:
+	 *       200:
+	 *         description: The request was successful.
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: '#/components/schemas/KeyResult'
+	 *       400:
+	 *         description: A problem with the input fields has occurred. Additional state information plus metadata may be available in the response body.
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: '#/components/schemas/InvalidRequest'
+	 *             example:
+	 *               error: InvalidRequest
+	 *       401:
+	 *         $ref: '#/components/schemas/UnauthorizedError'
+	 *       500:
+	 *         description: An internal error has occurred. Additional state information plus metadata may be available in the response body.
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: '#/components/schemas/InvalidRequest'
+	 *             example:
+	 *               error: Internal Error
+	 */
+	public async getKey(request: Request, response: Response) {
+		try {
+			const key = await new IdentityServiceStrategySetup(response.locals.customer.customerId).agent.getKey(
+				request.params.kid,
+				response.locals.customer
+			);
+			if (key) {
+				return response.status(StatusCodes.OK).json(key);
+			}
+			return response.status(StatusCodes.NOT_FOUND).json({
+				error: `Key with kid: ${request.params.kid} not found`,
+			});
+		} catch (error) {
+			return response.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
+				error: `${error}`,
+			});
+		}
+	}
+}
